+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added game sandboxing feature
+
+### Changed
+
+- Removed fractions displaying in components downloading progress bar
+- Moved to upgraded launcher SDK
+
 ## [3.4.1] - 12.04.2023
 
 ### Fixed